selfRightSpace: consistently use a double for calculating self width
This avoids going backwards and forwards between double and int, which was
causing some compiler warnings and unnecessary loss of precision.

+++ lib/common/render.h
@@ -137,7 +137,7 @@ extern "C" {
     RENDER_API void routesplinesterm(void);
     RENDER_API pointf* simpleSplineRoute (pointf, pointf, Ppoly_t, int*, int);
     RENDER_API pointf *routepolylines(path* pp, int* npoints);
-    RENDER_API int selfRightSpace (edge_t* e);
+    RENDER_API double selfRightSpace(edge_t *e);
     RENDER_API shape_kind shapeOf(node_t *);
     RENDER_API void shape_clip(node_t * n, pointf curve[4]);
     RENDER_API void make_simple_label (GVC_t * gvc, textlabel_t* rv);
